Fix unsigned int build for arm

+++ rts/Rendering/Textures/LegacyAtlasAlloc.cpp
@@ -167,7 +167,7 @@ int CLegacyAtlasAlloc::GetNumTexLevels() const
 {
 	RECOIL_DETAILED_TRACY_ZONE;
 	return std::min(
-		std::bit_width(static_cast<uint32_t>(GetMinDim())),
+		static_cast<int>(std::bit_width(static_cast<uint32_t>(GetMinDim()))),
 		numLevels
 	);
 }
